🔨 drop hideLegend
