+++ adminSiteClient/GrapherConfigGridEditor.tsx
@@ -1111,7 +1111,7 @@ export class GrapherConfigGridEditor extends React.Component<GrapherConfigGridEd
 
     async getFieldDefinitions() {
         const json = await fetch(
-            "https://files.ourworldindata.org/schemas/grapher-schema.009.json"
+            "https://files.ourworldindata.org/schemas/grapher-schema.010.json"
         ).then((response) => response.json())
         const fieldDescriptions = extractFieldDescriptionsFromSchema(json)
         runInAction(() => {
