+++ packages/@ourworldindata/grapher/src/lineCharts/LineChartThumbnail.tsx
@@ -263,7 +263,7 @@ export class LineChartThumbnail
         | undefined {
         if (!this.manager.showSeriesLabels) return undefined
 
-        const showSeriesLabels =
+        const showEntityNames =
             !this.manager.isDisplayedAlongsideComplementaryTable
 
         let labelCandidateSeries = this.chartState.series
@@ -298,7 +298,7 @@ export class LineChartThumbnail
                 const value = startPoint?.y ?? 0
 
                 const yPosition = this.outerBoundsVerticalAxis.place(value)
-                const label = showSeriesLabels
+                const label = showEntityNames
                     ? seriesName
                     : this.formatLabel(value)
 
@@ -324,8 +324,8 @@ export class LineChartThumbnail
         return new SimpleVerticalLabelsState(series, {
             fontSize: this.labelFontSize,
             fontWeight: 500,
-            maxWidth: showSeriesLabels ? 0.25 * this.bounds.width : undefined,
-            minSpacing: showSeriesLabels ? 5 : 2,
+            maxWidth: showEntityNames ? 0.25 * this.bounds.width : undefined,
+            minSpacing: showEntityNames ? 5 : 2,
             yRange: this.labelsRange,
             resolveCollision: (
                 s1: InitialSimpleLabelSeries,
