+++ packages/@ourworldindata/grapher/src/schema/defaultGrapherConfig.ts
@@ -4,7 +4,7 @@
 
 import { GrapherInterface } from "@ourworldindata/types"
 
-export const latestSchemaVersion = "009" as const
+export const latestSchemaVersion = "010" as const
 export const outdatedSchemaVersions = [
     "001",
     "002",
@@ -14,10 +14,11 @@ export const outdatedSchemaVersions = [
     "006",
     "007",
     "008",
+    "009",
 ] as const
 
 export const defaultGrapherConfig = {
-    $schema: "https://files.ourworldindata.org/schemas/grapher-schema.009.json",
+    $schema: "https://files.ourworldindata.org/schemas/grapher-schema.010.json",
     selectedEntityNames: [],
     focusedSeriesNames: [],
     excludedEntityNames: [],
@@ -67,7 +68,6 @@ export const defaultGrapherConfig = {
     },
     invertColorScheme: false,
     logo: "owid",
-    hideLegend: false,
     hideSeriesLabels: false,
     hideLogo: false,
     hideRelativeToggle: true,
