+++ packages/@ourworldindata/types/src/grapherTypes/GrapherTypes.ts
@@ -546,7 +546,6 @@ export interface GrapherInterface extends SortConfig {
     stackMode?: StackMode
 
     showNoDataArea?: boolean
-    hideLegend?: boolean
     hideSeriesLabels?: boolean
     logo?: LogoOption
     hideLogo?: boolean
@@ -682,7 +681,6 @@ export const grapherKeysToSerialize = [
     "addCountryMode",
     "stackMode",
     "showNoDataArea",
-    "hideLegend",
     "hideSeriesLabels",
     "logo",
     "hideLogo",
